+++ src/main/java/com/excilys/cdb/daos/ComputerDAO.java
@@ -90,7 +90,7 @@ public class ComputerDAO implements DAO<Computer> {
         statement = connection.prepareStatement(COUNT_COMPUTERS);
         rs = statement.executeQuery();
         if (rs.next()) {
-            double maxPage = rs.getInt(1) / page.getResultsPerPage();
+            double maxPage = (double) rs.getInt(1) / (double) page.getResultsPerPage();
             page.setMaxPage((int) Math.ceil(maxPage));
         }
